Migrate heap sort
- Move implementation to header file
- Add missing curly braces
- Update README
- Delete old heap_sort.cpp
