+++ C++/README.md
@@ -32,7 +32,7 @@ Implementation of well-known (and some rare) algorithms, in C++.
 * Sorting
   * [Bubble sort](include/algorithm/sorting/bubble_sort.hpp) :white_check_mark:
   * [Counting sort (stable)](include/algorithm/sorting/counting_sort.hpp) :white_check_mark:
-  * [Heap sort](source/sorting/heap_sort.cpp)
+  * [Heap sort](include/algorithm/sorting/heap_sort.hpp) :white_check_mark:
   * [Insertion sort](source/sorting/insertion_sort.cpp)
   * [Merge sort](source/sorting/merge_sort.cpp)
   * [Quick sort](source/sorting/quick_sort.cpp)
