+++ C++/test/algorithm/sorting/sorting.cpp
@@ -3,6 +3,7 @@
 #include "third_party/catch.hpp"
 #include "algorithm/sorting/bubble_sort.hpp"
 #include "algorithm/sorting/counting_sort.hpp"
+#include "algorithm/sorting/heap_sort.hpp"
 
 // Prototypes
 int generate_random_int(int, int);
@@ -18,7 +19,8 @@ TEST_CASE("Sort in ascending order", "[sorting]") {
 	// Sorting algorithms
 	vector<sorting_function> sorting_functions = {
 			bubble_sort,
-			counting_sort
+			counting_sort,
+			heap_sort
 	};
 
 	vector<int> original, algo_sorted, std_sorted;
@@ -41,7 +43,8 @@ TEST_CASE("Sort in descending order", "[sorting]") {
 	// Sorting algorithms
 	vector<sorting_function> sorting_functions = {
 			bubble_sort,
-			counting_sort
+			counting_sort,
+			heap_sort
 	};
 
 	vector<int> original, algo_sorted, std_sorted;
